src/product/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-ninja>=1.6.2",
    "django-cors-headers>=4.9.0",
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
]
